lelab-common/pyproject.toml
[project]
name = "lelab-common"
version = "0.1.0"
description = "Common utilities and shared components for lelab projects"
authors = []
requires-python = ">=3.12,<3.13"
readme = "README.md"
maintainers = []
dependencies = [
	"fastapi>=0.115.6",
	"pydantic>=2.10.4",
	"pydantic-settings>=2.7.0",
	"PyJWT>=2.8.0",
	"httpx>=0.27.0",
	"cachetools>=6.1.0",
	"starlette>=0.35.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
